Introduce simulator service
Which is now used only for starting the simulator and can later be extended to refresh it, etc.
